Fix Node.js version requirement and port validation
- Update native fetch API version from Node.js 25+ to Node.js 21+ in OAuth docs
- Fix OAuthConfigSchema server_port to require integers using z.number().int()

+++ packages/server-daemon/infrastructure/config/schemas.ts
@@ -18,7 +18,7 @@ export const LoggingConfigSchema = z.object({
 
 export const OAuthConfigSchema = z.object({
   redirect_uri: z.string().url().default('http://localhost:3000/callback'),
-  server_port: z.number().min(1).max(65535).default(3000),
+  server_port: z.number().int().min(1).max(65535).default(3000),
 });
 
 export const AppConfigSchema = z.object({
